+++ src/dis6/FastEntityStatePdu.cpp
@@ -719,7 +719,7 @@ int FastEntityStatePdu::getMarshalledSize() const
    marshalSize = marshalSize + 12 * 1;  // _marking
    marshalSize = marshalSize + 4;  // _capabilities
 
-   for(unsigned long idx=0; idx < _articulationParameters.size(); idx++)
+   for(unsigned long long idx=0; idx < _articulationParameters.size(); idx++)
    {
         ArticulationParameter listElement = _articulationParameters[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
