+++ src/dis7/CommentPdu.cpp
@@ -131,14 +131,14 @@ int CommentPdu::getMarshalledSize() const
    marshalSize = marshalSize + 4;  // _numberOfFixedDatumRecords
    marshalSize = marshalSize + 4;  // _numberOfVariableDatumRecords
 
-   for(unsigned long idx=0; idx < _fixedDatums.size(); idx++)
+   for(unsigned long long idx=0; idx < _fixedDatums.size(); idx++)
    {
         FixedDatum listElement = _fixedDatums[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
     }
 
 
-   for(unsigned long idx=0; idx < _variableDatums.size(); idx++)
+   for(unsigned long long idx=0; idx < _variableDatums.size(); idx++)
    {
         VariableDatum listElement = _variableDatums[idx];
         marshalSize = marshalSize + listElement.getMarshalledSize();
