+++ src/utils/DataStream.h
@@ -55,8 +55,8 @@ namespace DIS
       DataStream& operator <<(double c);
       DataStream& operator <<(int c);
       DataStream& operator <<(unsigned int c);
-      DataStream& operator <<(long c);
-      DataStream& operator <<(unsigned long c);
+      DataStream& operator <<(long long c);
+      DataStream& operator <<(unsigned long long c);
       DataStream& operator <<(unsigned short c);
       DataStream& operator <<(short c);
 
@@ -68,8 +68,8 @@ namespace DIS
       DataStream& operator >>(double& c);
       DataStream& operator >>(int& c);
       DataStream& operator >>(unsigned int& c);
-      DataStream& operator >>(long& c);
-      DataStream& operator >>(unsigned long& c);
+      DataStream& operator >>(long long& c);
+      DataStream& operator >>(unsigned long long& c);
       DataStream& operator >>(unsigned short& c);
       DataStream& operator >>(short& c);
